requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
pandas==2.2.3
numpy==2.2.4
python-multipart==0.0.20
openpyxl==3.1.5
Pillow==11.1.0
matplotlib==3.10.1
python-docx==1.1.2
python-dotenv==1.0.1
openai>=1.0.0
